+++ CHANGELOG.md
@@ -9,7 +9,7 @@ All notable changes to this project will be documented in this file.
 
 
 ### Changed
-- Added ROW_ID and EXPANDED_ID to the ind cqc columns class and updated references to the objects.
+- Added row_id and expanded_id to the ind cqc columns class and updated references to the objects.
 
 ### Fixed